Modified to work with the jedi Lighting 2in1
This version of the script remembers whether the light is on and when it has been turned off the last time, so that it can conclude whether the lamp is currently in warm white or cold white mode.

The light supported is the "http://www.jedi-light.com/de_DE/page/2-in-1". Currently the color state is posted per mqtt to "stat/keller_decke_2/CTEMP" in mireds hard coded. Commands to "cmnd/keller_decke_2/CTEMP" are converted in correct possible color values automically and the sonoff turns the light on and off to reach the desired color mode.