Fixed a bug and saved code at once
Now if a CTEMP signal is sent to the lamp, the lamp remembers this. The color is then enforced if the lamp is turned on via MQTT power on commands (which are likely to be sent when using sth. like home assistant)
If the user chooses to toggle the lamp via the button and has set the correct button setting (SetOption11 = 1), then the user may control the color of the lamp via the normal double toggleing.

This will remove the issue of home assistant turning the lamp on at random as it resends color commands on its own once in a while